Prep for next dev cycle, and sync with MAME 0.114u1
git-svn-id: svn://messdev.no-ip.org/mess@269 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mame/machine/vacfdisp.h
@@ -1,5 +1,3 @@
-// Any fixes for this driver should be forwarded to AGEMAME HQ (http://www.mameworld.net/agemame/)
-
 #ifndef VACFDISP
 #define VACFDISP
 
@@ -17,9 +15,9 @@ void	vfd_shift_data(int id, int data);	// clock in a bit of data
 
 int		vfd_newdata(   int id, int data);	// clock in 8 bits of data (scorpion2 vfd interface)
 
-UINT16	*vfd_get_segments(int id);			// get current segments displayed
-UINT16  *vfd_set_outputs(int id);			// convert segments to standard for display
-UINT16  *vfd_get_outputs(int id);			// get converted segments
+UINT32	*vfd_get_segments(int id);			// get current segments displayed
+UINT32  *vfd_set_outputs(int id);			// convert segments to standard for display
+UINT32  *vfd_get_outputs(int id);			// get converted segments
 
 char	*vfd_get_string( int id);			// get current string   displayed (not as accurate)
